+++ CHANGES.rst
@@ -32,6 +32,7 @@ Release 3.2.5 (2016-7-13)
 - Fix not showing timeseries for timeseries of nested assets.
 
 - Fix timeseries of parent assets not having thresholds.
+- Force login when requesting favourite unauthenticated.
 
 
 Release 3.2.4 (2016-6-20)